refactor: extract admin_api to seperate file
